infra/airflow/dags/requirements.txt
celery==5.3.1
werkzeug==2.2.3
SQLAlchemy==1.4.51
psycopg2-binary==2.9.9
statsmodels==0.14.2
langchain==0.3.26
langchain-core==0.3.80
langchain-community==0.3.27
langchain-openai==0.3.25
langgraph-checkpoint-postgres==2.0.22
Cython==0.29.35
cmdstanpy==1.1.0
numpy==1.26.4
pandas==1.5.2
setuptools==78.1.1
setuptools-git==1.2
python-dateutil>=2.9.0
tqdm==4.66.1
wheel==0.38.1
prophet==1.1.5
boto3==1.33.13
python-dotenv==0.21.0
markupsafe==2.1.3
slack==0.0.2
slackclient==2.9.4
pkutils==3.0.2
scikit-learn==1.5.2
pycryptodome==3.23.0
pycryptodomex==3.23.0
pmdarima==2.0.3
xhtml2pdf==0.2.15
pyodbc==5.1.0
pymsteams==0.2.2
detect-delimiter==0.1.1
thrift-sasl==0.4.3
sqlparse==0.5.0
SQLGlot==26.29.0
sql_metadata==2.11.0
pandasql==0.7.3
cryptography==44.0.1
pip-system-certs==4.0
sendgrid==6.11.0
selenium
pyatlan==6.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
spacy==3.8.0
sentence-transformers==4.0.1
fuzzywuzzy==0.18.0
RapidFuzz==3.12.2
gensim==4.3.3
opencv-python-headless
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.6.0
transformers>=4.50.0
typing_extensions>=4.10.0
awscrt==0.27.6
crccheck==1.3.1
google-crc32c==1.5.0
ijson==3.4.0
azure-storage-blob==12.21.0